Сombine getVideoAsync, addVideo and enterVideoName actions
